requirements.txt
asgiref==3.5.0
Django==3.1.6
django-tinymce4-lite==1.8.0
jsmin==3.0.1
Pillow==9.1.0
python-http-client==3.3.7
pytz==2022.1
sendgrid==6.9.7
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
